Remove obsolete asset naming strategies, interfaces, and test classes; consolidate namespaces and eliminate `FluentResults` usage throughout.
