+++ src/mlrb/MyLittleRangeBook/Config/BootstrapFuncs.cs
@@ -1,6 +1,4 @@
 ﻿using System.Text.Json.Nodes;
-using FluentResults;
-using Serilog;
 
 namespace MyLittleRangeBook.Config
 {
