+++ src/mlrb/MyLittleRangeBook/GlobalUsings.cs
@@ -1,0 +1,2 @@
+﻿global using FluentResults;
+global using Serilog;
