+++ src/mlrb/MyLittleRangeBook/RangeEventAssets/DuplicateFitFileNameError.cs
@@ -1,4 +1,4 @@
-﻿namespace MyLittleRangeBook.Persistence
+﻿namespace MyLittleRangeBook.RangeEventAssets
 {
     /// <summary>
     ///     Thrown when attempting to insert a FIT file with a name that already exists in the database - this indicates that
