+++ src/mlrb/MyLittleRangeBook/RangeEvents/ISimpleRangeEventHelper.cs
@@ -1,6 +1,4 @@
-﻿using FluentResults;
-
-namespace MyLittleRangeBook.RangeEvents
+﻿namespace MyLittleRangeBook.RangeEvents
 {
     /// <summary>
     ///     Helper interface for getting data about firearms and ranges when trying to create a range event.
