+++ src/mlrb/MyLittleRangeBook/RangeEvents/SqliteSimpleRangeEventService.cs
@@ -1,6 +1,5 @@
 ﻿using System.Data;
 using Dapper;
-using FluentResults;
 using Microsoft.Data.Sqlite;
 using MyLittleRangeBook.Models;
 using MyLittleRangeBook.RangeEvent;
